### 2024-02-13 #6351 @editorialbot
[PRE REVIEW]: MOTrainer: Distributed Measurement Operator Trainer for Data Assimilation Applications
**Submitting author:** @rogerkuou (<a href="http://orcid.org/0000-0002-5373-5209">Ou Ku</a>)
**Repository:** https://github.com/VegeWaterDynamics/motrainer
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
