### 2024-03-16 #6351 @arfon
@rogerkuou - thanks for your submission to JOSS. This submission has passed an initial scope review. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
