### 2024-03-24 #6351 @arfon
:wave: @srmnitc – would you be willing to edit this submission for JOSS? It feels _somewhat_ close(ish!) to 'research data management'.
