### 2024-04-05 #6351 @KwickSilver
Hi Sarath,

The paper seems interesting, and I'm eager to contribute to the review
process. Could you please include me as a reviewer as well?

Thank you,
Shantanu

On Fri, 5 Apr 2024 at 04:01, The Open Journals editorial robot <
***@***.***> wrote:

> @abhishektiwari <https://github.com/abhishektiwari> added to the
> reviewers list!
>
