### 2024-04-07 #6351 @d-eggert
@srmnitc The publication looks interesting, but I have no time available at the moment, sorry.
