### 2024-04-16 #6591 @srmnitc
> Hi Sarath, Apologies for my delayed response; I haven't been able to start reviewing the journal yet. I plan to begin tomorrow and aim to complete it by the end of this month. Does this timeline work for you? Thanks, Shantanu
> […](#)
> On Mon, 8 Apr 2024 at 04:47, Sarath Menon ***@***.***> wrote: 👋🏼 @rogerkuou <https://github.com/rogerkuou> @abhishektiwari <https://github.com/abhishektiwari> @KwickSilver <https://github.com/KwickSilver> this is the review thread for the paper. All of our communications will happen here from now on. As you might know, as a reviewer, the first step is to create a checklist for your review by entering @editorialbot <https://github.com/editorialbot> generate my checklist as the top of a new comment in this thread. These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines. The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention openjournals/joss-reviews#REVIEW_NUMBER so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package. We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time. Please feel free to ping me ***@***.*** <https://github.com/srmnitc> ) if you have any questions/concerns, thanks again for the submission, and for thr reviews — Reply to this email directly, view it on GitHub <[#6591 (comment)](https://github.com/openjournals/joss-reviews/issues/6591#issuecomment-2042538883)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ABCUHHJ5AU25RZTRQPVGMKDY4J7VBAVCNFSM6AAAAABF4RFJNSVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDANBSGUZTQOBYGM> . You are receiving this because you were mentioned.Message ID: ***@***.***>

@KwickSilver that is fine, thanks for the update!
