### 2024-04-28 VegeWaterDynamics/motrainer#118 @abhishektiwari
[JOSS] Installation instruction: Could not build wheels for h5py
Package installation on Ubuntu runs into following error,

```
ERROR: Could not build wheels for h5py, which is required to install pyproject.toml-based projects
```

Please update the installation instruction for dependencies. Package requires `h5py` which in turn needs `hdf5` library installed on the OS. For Linux/Ubunutu one requires to install `libhdf5-dev`, so provide specific instructions for Mac, Linux, and Windows.

```
sudo apt-get install libhdf5-dev
```

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6591)
