### 2024-04-28 VegeWaterDynamics/motrainer#119 @abhishektiwari
[JOSS] Example dataset generator Jupyter Notebook
At multiple places documentation suggest that

> The example dataset ./example1_data.zarr/ can be generated using this [Jupyter Notebook](https://vegewaterdynamics.github.io/motrainer/notebooks/example_daskml/).

I was unable to find the code snippet or script to generate example dataset.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6591)
