### 2024-04-28 VegeWaterDynamics/motrainer#120 @abhishektiwari
[JOSS] Add community guidelines
Please Community guidelines to either `README.md` or seperate `CONTRIBUTING.md`  for any wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6591)
