### 2024-04-28 #6591 @abhishektiwari
@rogerkuou Initial feedback,

1. Please update installation instruction as raised [here](https://github.com/VegeWaterDynamics/motrainer/issues/118)
2. Please provide details on example data generation script [here](https://github.com/VegeWaterDynamics/motrainer/issues/119)
3. Add community/contributing guidelines as mentioned [here](https://github.com/VegeWaterDynamics/motrainer/issues/120)

Once you have fixed these, let us know in comments here so I can verify functional claims.
