### 2024-05-10 5b4b02b
adapt to new keras changes to fix unit testing
