### 2024-05-14 8b737e5
remove the model saving since TensorFlow graph execution issue has been solved
