+++ pyproject.toml
@@ -5,12 +5,12 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "motrainer"
 version = "0.1.3"
-requires-python = ">=3.10, <3.12"
+requires-python = ">=3.10"
 dependencies = [
     "dask[complete]",
     "ipykernel",
-    "numpy<1.24.0",
-    "pandas==2.0.3",
+    "numpy",
+    "pandas",
     "scikit-learn",
     "scikit-optimize",
     "scipy",
