+++ docs/setup.md
@@ -1,5 +1,7 @@
 # Installation
 
+Python version `>=3.10` is required to install MOTrainer.
+
 MOTrainer can be installed from PyPI:
 
 ```sh
@@ -14,8 +16,22 @@ cd motrainer
 pip install .
 ```
 
-Note that Python version `>=3.10` and `<3.12` is required to install MOTrainer.
-
 ## Tips
 
 We strongly recommend installing separately from your default Python envrionment. E.g. you can use enviroment manager e.g. [mamba](https://mamba.readthedocs.io/en/latest/mamba-installation.html) to create separate environment.
+
+## Troubleshooting
+
+### Error: Could not build wheels for h5py
+
+The following error may occur when installing `h5py` on Ubuntu system:
+
+```output
+ERROR: Could not build wheels for h5py, which is required to install pyproject.toml-based project
+```
+
+This can be resolved by installing the `libhdf5-dev` package:
+
+```bash
+sudo apt-get install libhdf5-dev
+```
