### 2024-05-16 #6591 @rogerkuou
Hi @srmnitc, thanks for checking! And apologies for the delay.
I'm still in progress working on the comments from @abhishektiwari. Two PRs already ready but we are having a hiccup with unit test taking too long to run for MacOS (For example, [this Action)](https://github.com/VegeWaterDynamics/motrainer/actions/runs/9108438187/job/25039203129). This can be caused by an updated version of TensorFlow. I am working on this. Will post it here once it's solved.
