### 2024-05-16 VegeWaterDynamics/motrainer#124 @rogerkuou
MacOS Github Action takes too long to run
MacOS Github Action does not finish in hours. It is stuck with the unit test. This does not happen for Windows and Ubuntu.
See this run: https://github.com/VegeWaterDynamics/motrainer/actions/runs/9095957227

Temporally disabled MacOS test because of this.

Some observations:
Tests without need to import `pandas works`. See this run: https://github.com/VegeWaterDynamics/motrainer/actions/runs/9111642386

If exclude `test_util.py`, others still does not work. See this run: https://github.com/VegeWaterDynamics/motrainer/actions/runs/9112310974

Constraining `pandas`  version or not does not matter. See the following two runs:
- https://github.com/VegeWaterDynamics/motrainer/actions/runs/9112310974
- https://github.com/VegeWaterDynamics/motrainer/actions/runs/9112375686
