### 2024-05-21 8e0907a
also test python 3.12
