### 2024-05-21 2dc5e5d
Update .github/workflows/build.yml
Co-authored-by: SarahAlidoost <55081872+SarahAlidoost@users.noreply.github.com>
