+++ .github/workflows/build.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ['ubuntu-latest', 'windows-latest']
+        os: ['ubuntu-latest', 'windows-latest']   # 'macos-latest' is not included, see issue 124
         python-version: ['3.10', '3.11', '3.12']
     steps:
       - uses: actions/checkout@v3
