### 2024-05-29 aaaa583
constraint scikit-learn version
