### 2024-05-29 VegeWaterDynamics/motrainer#120 @rogerkuou
Hi @abhishektiwari, there is a [CONTRIBUTING](https://vegewaterdynamics.github.io/motrainer/CONTRIBUTING/) page in the documentation, but I agree it is not obvious enough.
In PR#125 I added a link of CONTRIBUTING page the to the README file.
