### 2024-05-29 #6591 @rogerkuou
Hi @srmnitc and @abhishektiwari
Sorry for the long waiting!
I have stressed the three comments raised by @abhishektiwari in separate PRs. See the reply in the following three issues:

- https://github.com/VegeWaterDynamics/motrainer/issues/118
- https://github.com/VegeWaterDynamics/motrainer/issues/119
- https://github.com/VegeWaterDynamics/motrainer/issues/120

When working on these issues, I found the GitHub Action on MacOS takes forever to execute. I did not duplicate this on my colleagues MacOS PC. I have temporally disabled the test for MacOS and looking into this issue: https://github.com/VegeWaterDynamics/motrainer/issues/124

Meanwhile, do you think you can continue reviewing with this?

Thanks for taking the time to review! Let me know if there are more issues coming up.
