### 2024-06-01 #6591 @abhishektiwari
@rogerkuou

Re: MacOS builds - I think that is something to do with ARM/M1 architecture of OS. Let me review error logs and see if I can help you there.

That said you are probably only JOSS paper in so far who is doing multi-platform build. So kudos for supporting Linux and Windows build from get go.
