### 2024-06-09 #6591 @abhishektiwari
@rogerkuou Thanks for updating documentation. Overall looks good.

For some reason, example notebook for _Prallely training DNN with Tensorflow_ shows [error on page load](https://vegewaterdynamics.github.io/motrainer/notebooks/example_dnn/).

@srmnitc Review checklist is complete from my side. Please let me know if any questions.
