### 2024-06-10 #6591 @srmnitc
@abhishektiwari thanks a lot for your efforts, could you please confirm that you recommend publication of this package in JOSS?
