### 2024-06-11 #6591 @abhishektiwari
@srmnitc Yes, I recommend publication of this package in JOSS.
