### 2024-06-19 VegeWaterDynamics/motrainer#124 @SarahAlidoost
@rogerkuou Looking at the history of changes, it seems that the tests have been passed on macos before. Unfortunately, the logs of previous actions are not available anymore. Otherwise, we could compare the list of dependencies installed by GA before and after.

Also some dependencies might behave differently on macos with arch arm64 and x86. You can also do a test with `macos-12` that works on arch x86. [Here](https://github.com/SDCCA/DGL-PTM/blob/d7cabe37503e91bc628b9b9f435b9dd93b5b15cd/.github/workflows/build_macos_x64.yml) is an example workflow. Note the command `arch -x86_64` before python commands.
