### 2024-06-28 #6591 @srmnitc
> @srmnitc Sure thing! I'm going to be away at a conference the week of the 8th-12th, but I'll get as much done as I can over this upcoming week. Is it okay if I potentially take up through the following week of the 15th-19th?

That would be ok, thanks!
