### 2024-07-16 #6591 @WarmCyan
@rogerkuou and @srmnitc I've completed the review, overall great work! It's an interesting library and the repository seems to be quite well put together.

I have a couple very minor recommendations:
* https://github.com/VegeWaterDynamics/motrainer/issues/126
* https://github.com/VegeWaterDynamics/motrainer/issues/127
