+++ CITATION.cff
@@ -1,7 +1,7 @@
 # This CITATION.cff file was generated with cffinit.
 # Visit https://bit.ly/cffinit to generate yours today!
 
-cff-version: 1.2.0
+cff-version: 0.1.4
 title: MOTrainer
 message: >-
   Distributed Measurement Operator Trainer for Data
