### 2024-08-01 #6591 @rogerkuou
> @rogerkuou Thanks for updating documentation. Overall looks good.
>
> For some reason, example notebook for _Prallely training DNN with Tensorflow_ shows [error on page load](https://vegewaterdynamics.github.io/motrainer/notebooks/example_dnn/).
>
> @srmnitc Review checklist is complete from my side. Please let me know if any questions.

And I realized I missed the comment from @abhishektiwari. The documentation page has been fixed. Now the page should work: https://vegewaterdynamics.github.io/motrainer/notebooks/example_dnn/
