+++ mkdocs.yml
@@ -13,8 +13,8 @@ nav:
   - Examples: 
     - Covert a nested DataFrame to a Dataset: notebooks/example_read_from_one_df.ipynb
     - Align multiple DataFrames to a Dataset: notebooks/example_read_from_multiple_df.ipynb
-    - Prallely training sklearn models with dask-ml: notebooks/example_daskml.ipynb
-    - Prallely training DNN with Tensorflow: notebooks/example_dnn.ipynb
+    - Parallelly training sklearn models with dask-ml: notebooks/example_daskml.ipynb
+    - Parallelly training DNN with Tensorflow: notebooks/example_dnn.ipynb
   - Conributing guide: CONTRIBUTING.md
   - Change Log: CHANGELOG.md
   - API Reference: api_reference.md
