### 2024-08-19 VegeWaterDynamics/motrainer#130 @srmnitc
Changes to paper
This is part of the review at https://github.com/openjournals/joss-reviews/issues/6591

I suggest some small changes to the paper:
- add ORCIDs for all authors
- line 9: could be strengthened with a citation
- line 16-17: the sentence is not clear to me. Could you please check for correctness and rephrase it?
- line 30, line 34: there are abbreviation CV and LSM introduced which are not used again.
- line 65: reference DOI should not have https://doi.org prefix
- line 69: reference DOI should not have https://doi.org prefix

Thanks a lot!
