+++ paper/paper.bib
@@ -6,7 +6,7 @@
   pages = {113116},
   year = {2022},
   issn = {0034-4257},
-  doi = {https://doi.org/10.1016/j.rse.2022.113116},
+  doi = {10.1016/j.rse.2022.113116},
   url = {https://www.sciencedirect.com/science/article/pii/S0034425722002309},
   author = {Xu Shan and Susan Steele-Dunne and Manuel Huber and Sebastian Hahn and Wolfgang Wagner and Bertrand Bonan and Clement Albergel and Jean-Christophe Calvet and Ou Ku and Sonja Georgievska},
   keywords = {ASCAT, Scatterometry, Radar, Vegetation, Land surface model, Machine learning, Deep Neural Network, Plant water dynamics, Soil moisture},
@@ -31,7 +31,7 @@
   pages = {153-165},
   year = {2015},
   issn = {0034-4257},
-  doi = {https://doi.org/10.1016/j.rse.2015.09.009},
+  doi = {doi.org/10.1016/j.rse.2015.09.009},
   url = {https://www.sciencedirect.com/science/article/pii/S0034425715301322},
   author = {Yuan Xue and Barton A. Forman},
   keywords = {Sensitivity analysis, Machine learning, Brightness temperature, Snow, Data assimilation},
@@ -80,4 +80,69 @@
   booktitle={SciPy},
   year={2015},
   url={https://api.semanticscholar.org/CorpusID:63554230}
+}
+
+@article{REICHLE20081411,
+title = {Data assimilation methods in the Earth sciences},
+journal = {Advances in Water Resources},
+volume = {31},
+number = {11},
+pages = {1411-1418},
+year = {2008},
+note = {Hydrologic Remote Sensing},
+issn = {0309-1708},
+doi = {10.1016/j.advwatres.2008.01.001},
+url = {https://www.sciencedirect.com/science/article/pii/S0309170808000043},
+author = {Rolf H. Reichle},
+keywords = {Data assimilation, Remote sensing, Land surface hydrology, Variational methods, Kalman filter},
+abstract = {Although remote sensing data are often plentiful, they do not usually satisfy the users’ needs directly. Data assimilation is required to extract information about geophysical fields of interest from the remote sensing observations and to make the data more accessible to users. Remote sensing may provide, for example, measurements of surface soil moisture, snow water equivalent, snow cover, or land surface (skin) temperature. Data assimilation can then be used to estimate variables that are not directly observed from space but are needed for applications, for instance root zone soil moisture or land surface fluxes. The paper provides a brief introduction to modern data assimilation methods in the Earth sciences, their applications, and pertinent research questions. Our general overview is readily accessible to hydrologic remote sensing scientists. Within the general context of Earth science data assimilation, we point to examples of the assimilation of remotely sensed observations in land surface hydrology.}
+}
+
+@article{Carrassi2018,
+  title={Data assimilation in the geosciences: An overview of methods, issues, and perspectives},
+  author={Carrassi, Alberto and Bocquet, Marc and Bertino, Laurent and Evensen, Geir},
+  journal={Wiley Interdisciplinary Reviews: Climate Change},
+  volume={9},
+  number={5},
+  pages={e535},
+  year={2018},
+  publisher={Wiley Online Library},
+  doi={10.1002/wcc.535}
+}
+
+@ARTICLE{Evensen2009,
+  author={Evensen, Geir},
+  journal={IEEE Control Systems Magazine}, 
+  title={The ensemble Kalman filter for combined state and parameter estimation}, 
+  year={2009},
+  volume={29},
+  number={3},
+  pages={83-104},
[…]
