+++ paper/paper.bib
@@ -79,7 +79,8 @@
   author={Matthew Rocklin},
   booktitle={SciPy},
   year={2015},
-  url={https://api.semanticscholar.org/CorpusID:63554230}
+  url={https://api.semanticscholar.org/CorpusID:63554230},
+  doi={10.25080/majora-7b98e3ed-013}
 }
 
 @article{REICHLE20081411,
