+++ paper/paper.bib
@@ -31,7 +31,7 @@
   pages = {153-165},
   year = {2015},
   issn = {0034-4257},
-  doi = {doi.org/10.1016/j.rse.2015.09.009},
+  doi = {10.1016/j.rse.2015.09.009},
   url = {https://www.sciencedirect.com/science/article/pii/S0034425715301322},
   author = {Yuan Xue and Barton A. Forman},
   keywords = {Sensitivity analysis, Machine learning, Brightness temperature, Snow, Data assimilation},
