### 2024-08-20 #6591 @rogerkuou
@srmnitc Thanks for the comments and suggestions. I made some revise in PR: https://github.com/VegeWaterDynamics/motrainer/pull/131 to address the comments you made.

Below are the required information:

- Version number: [v0.1.6](https://github.com/VegeWaterDynamics/motrainer/releases/tag/v0.1.6)
- Software DOI: [10.5281/zenodo.7540442](https://zenodo.org/doi/10.5281/zenodo.7540442) (This is a general DOI for all versions, I hope it's okay. The doi specific for v0.1.6 is [10.5281/zenodo.13348729](https://doi.org/10.5281/zenodo.13348729))

If you find there is anything else that needs to be changed/improved, just let me know. And please take your time. It can also be done after you are back in the office in September.
