### 2024-08-20 a800634
change references for jackknife
