+++ paper/paper.bib
@@ -52,16 +52,17 @@
   eprint = {https://doi.org/10.1080/02723646.2016.1236606}
   }
 
-@book{mccuen1998hydrologic,
-  title={Hydrologic Analysis and Design},
-  author={McCuen, R.H.},
-  isbn={9780131349582},
-  lccn={97044779},
-  series={Hewlett Packard Professional Books},
-  url={https://books.google.com.mt/books?id=qPdRAAAAMAAJ},
-  year={1998},
-  publisher={Prentice Hall}
-  }
+@book{Efron1982,
+  title     = {The Jackknife, the Bootstrap and Other Resampling Plans},
+  author    = {Bradley Efron},
+  year      = {1982},
+  series    = {CBMS-NSF Regional Conference Series in Applied Mathematics},
+  volume    = {38},
+  publisher = {SIAM},
+  address   = {Philadelphia},
+  doi       = {10.1137/1.9781611970319},
+  url       = {http://dx.doi.org/10.1137/1.9781611970319}
+}
 
 @article{Hoyer_xarray_N-D_labeled_2017,
   author = {Hoyer, Stephan and Joseph, Hamman},
