+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "motrainer"
-version = "0.1.6"
+version = "0.1.7"
 requires-python = ">=3.10"
 dependencies = [
     "dask[complete]",
