### 2024-08-20 #6591 @rogerkuou
> @srmnitc Thanks for the comments and suggestions. I made some revise in PR: [VegeWaterDynamics/motrainer#131](https://github.com/VegeWaterDynamics/motrainer/pull/131) to address the comments you made. The newest version (v0.1.6) including the changes has been released.
>
> There is one citation missing DOI. It is a reference book for the "Jackknife" method. Unfortunately, it seems to have no DOI available.
>
> Below are the required information:
>
> * Version number: [v0.1.6](https://github.com/VegeWaterDynamics/motrainer/releases/tag/v0.1.6)
> * Software DOI: [10.5281/zenodo.7540442](https://zenodo.org/doi/10.5281/zenodo.7540442) (This is a general DOI for all versions, I hope it's okay. The doi specific for v0.1.6 is [10.5281/zenodo.13348729](https://doi.org/10.5281/zenodo.13348729))
>
> If you find there is anything else that needs to be changed/improved, just let me know. And please take your time. It can also be done after you are back in the office in September.

Hi @srmnitc, follow my previous reply, I checked the reference again and decided to cite the origin of Jackknife method instead. Now the DOI problem should have been solved.

Below is the updated info:

- Version number: [v0.1.7](https://github.com/VegeWaterDynamics/motrainer/releases/tag/v0.1.7)
- Software DOI for all versions: [10.5281/zenodo.7540442](https://zenodo.org/doi/10.5281/zenodo.7540442)
- Software DOI for v0.1.7: [10.5281/zenodo.13349186](https://doi.org/10.5281/zenodo.13349186)

My apology for spamming. Please take your time and see if this suffices.
