### 2024-08-20 #6591 @srmnitc
> @rogerkuou I see that you also have released v1.0.7, can you please confirm it is indeed v1.0.6 that should be mentioned?

Ah, thanks for clarifying, I did not see your message.
