### 2024-08-20 #6591 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
