### 2024-08-20 #6591 @rogerkuou
> @editorialbot set v1.0.7 as version

Hi @srmnitc sorry for bothering you again, but seems you set the version to v1.0.7. The version I released is [v0.1.7](https://github.com/VegeWaterDynamics/motrainer/releases/tag/v0.1.7). I should have noticed this earlier.

Maybe it can be easily changed? I am not sure if I can use the set version command from editorialbot.

And again, thanks a lot for the editorial work!
