### 2024-08-22 #6591 @crvernon
:wave: @rogerkuou - I need you to address the following before I can move to accept this one for publication.

- [ ] The title of your Zenodo archive should match the title of your paper exactly.  Please edit the metadata in your Zenodo archive to use the title of your paper.  No need to conduct another release.

In the paper:
- [ ] Please ensure that capitalization is maintained where necessary in your references.  For example in LINE 72, the "e" in "earth" should be capitalized.  You can do this in your bib file by using curly brackets around text that you wish to maintain the formatting of.

Let me know once you address these and I will move to accept the paper.
