### 2024-08-22 a0dfb6e
fix cases in references
