+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{shan:2022,
-  title = {Towards constraining soil and vegetation dynamics in land surface models: Modeling ASCAT backscatter incidence-angle dependence with a Deep Neural Network},
+  title = {{Towards constraining soil and vegetation dynamics in land surface models: Modeling ASCAT backscatter incidence-angle dependence with a Deep Neural Network}},
   journal = {Remote Sensing of Environment},
   volume = {279},
   pages = {113116},
@@ -18,14 +18,14 @@
   year = {2014},
   month = {06},
   pages = {1-11},
-  title = {Using a Support Vector Machine and a Land Surface Model to Estimate Large-Scale Passive Microwave Brightness Temperatures Over Snow-Covered Land in North America},
+  title = {{Using a Support Vector Machine and a Land Surface Model to Estimate Large-Scale Passive Microwave Brightness Temperatures Over Snow-Covered Land in North America}},
   volume = {8},
   journal = {IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing},
   doi = {10.1109/JSTARS.2014.2325780}
   }
 
 @article{XUE:2015,
-  title = {Comparison of passive microwave brightness temperature prediction sensitivities over snow-covered land in North America using machine learning algorithms and the Advanced Microwave Scanning Radiometer},
+  title = {{Comparison of passive microwave brightness temperature prediction sensitivities over snow-covered land in North America using machine learning algorithms and the Advanced Microwave Scanning Radiometer}},
   journal = {Remote Sensing of Environment},
   volume = {170},
   pages = {153-165},
@@ -40,7 +40,7 @@
 
 @article{Forman:2017,
   author = {Barton A. Forman and Yuan Xue},
-  title = {Machine learning predictions of passive microwave brightness temperature over snow-covered land using the special sensor microwave imager (SSM/I)},
+  title = {{Machine learning predictions of passive microwave brightness temperature over snow-covered land using the special sensor microwave imager (SSM/I)}},
   journal = {Physical Geography},
   volume = {38},
   number = {2},
@@ -53,7 +53,7 @@
   }
 
 @book{Efron1982,
-  title     = {The Jackknife, the Bootstrap and Other Resampling Plans},
+  title     = {{The Jackknife, the Bootstrap and Other Resampling Plans}},
   author    = {Bradley Efron},
   year      = {1982},
   series    = {CBMS-NSF Regional Conference Series in Applied Mathematics},
@@ -76,7 +76,7 @@
   }
 
 @inproceedings{Rocklin2015DaskPC,
-  title={Dask: Parallel Computation with Blocked algorithms and Task Scheduling},
+  title={{Dask: Parallel Computation with Blocked algorithms and Task Scheduling}},
   author={Matthew Rocklin},
   booktitle={SciPy},
   year={2015},
@@ -85,7 +85,7 @@
 }
 
 @article{REICHLE20081411,
-title = {Data assimilation methods in the Earth sciences},
+title = {{Data assimilation methods in the Earth sciences}},
 journal = {Advances in Water Resources},
 volume = {31},
 number = {11},
@@ -115,7 +115,7 @@ abstract = {Although remote sensing data are often plentiful, they do not usuall
 @ARTICLE{Evensen2009,
   author={Evensen, Geir},
   journal={IEEE Control Systems Magazine}, 
-  title={The ensemble Kalman filter for combined state and parameter estimation}, 
+  title={{The ensemble Kalman filter for combined state and parameter estimation}}, 
   year={2009},
   volume={29},
[…]
