+++ paper/paper.bib
@@ -101,7 +101,7 @@ abstract = {Although remote sensing data are often plentiful, they do not usuall
 }
 
 @article{Carrassi2018,
-  title={Data assimilation in the geosciences: An overview of methods, issues, and perspectives},
+  title={{Data assimilation in the geosciences: An overview of methods, issues, and perspectives}},
   author={Carrassi, Alberto and Bocquet, Marc and Bertino, Laurent and Evensen, Geir},
   journal={Wiley Interdisciplinary Reviews: Climate Change},
   volume={9},
