### 2024-08-22 #6591 @rogerkuou
Hi @crvernon, thanks a lot for the review!

The [Zenodo repo](https://zenodo.org/records/13349186) title and [metadata file](https://github.com/VegeWaterDynamics/motrainer/blob/main/zenodo.json). Now the title strictly matches the JOSS paper

The reference capitalization has been solved by this PR: https://github.com/VegeWaterDynamics/motrainer/pull/132
